all kindz of gainz - tomorrow only js programming left for "search" - generic template will be used, as for create/update dialogs - for each of those it will be generic.

+++ webclient/controller/test.py
@@ -4,8 +4,10 @@ Created on Oct 10, 2013
 
 @author:  Edis Sehalic (edis.sehalic@gmail.com)
 '''
-from webclient import handler
- 
+import inspect
+
+from webclient import handler 
+
 from app import ndb, memcache
 from app.srv import io
 
@@ -18,18 +20,30 @@ class Reset(handler.Angular):
       from app import opt
     
       models = [srv.auth.Domain, srv.log.Record, srv.setup.Configuration, srv.auth.User, srv.rule.DomainRole, srv.rule.DomainUser,
+                srv.setup.Configuration, srv.nav.Widget, srv.event.Action, srv.notify.Template,
                 domain.business.Company, domain.business.CompanyContent, domain.marketing.Catalog,
                 domain.marketing.CatalogImage, domain.marketing.CatalogPricetag, domain.product.Content,
                 domain.product.Instance, domain.product.InventoryAdjustment, domain.product.InventoryLog,
                 domain.product.Variant, domain.product.Template, opt.buyer.Address, opt.buyer.Collection, 
                 opt.misc.Content, opt.misc.ProductCategory, opt.misc.SupportRequest]
- 
-      for mod in models:
-          ndb.delete_multi(mod.query().fetch(keys_only=True))
-          
       
-      memcache.flush_all()
-      return {'models' : [f.__name__ for f in models]}
+      if self.request.get('delete'):
+        for mod in models:
+            ndb.delete_multi(mod.query().fetch(keys_only=True))
+             
+        memcache.flush_all()
+        
+      paths = {}
+      
+      for mod in models:
+          path = inspect.getfile(mod)
+          pathsplit = path.split('/app/')
+          if pathsplit and pathsplit[1]:
+             modelpath = '%s.%s' % (pathsplit[1].split('.')[0], mod.__name__)
+             modelpath = modelpath.replace('/', '.')
+             paths[mod.get_kind()] = modelpath
+        
+      return {'models' : [f.__name__ for f in models], 'paths' : paths}
 
 class Submitter(handler.Angular):
